refactor: speed up hcl2cdk tests
we now parallelize as much as possible and split the tests up into several files
