+++ packages/@cdktf/hcl2cdk/.gitignore
@@ -3,3 +3,4 @@
 tsconfig.tsbuildinfo
 !tsconfig.json
 cdktf-convert-*
+cdktf-provider*
