+++ packages/@cdktf/hcl2cdk/package.json
@@ -47,6 +47,7 @@
     "@types/jest": "^26.0.24",
     "@types/node": "^14.18.30",
     "@types/reserved-words": "^0.1.0",
+    "deepmerge": "^4.2.2",
     "jest": "^26.6.3",
     "ts-jest": "^26.5.6",
     "typescript": "^3.9.10"
