+++ packages/@cdktf/hcl2cdk/test/locals.test.ts
@@ -1,0 +1,40 @@
+import { testCase, Synth } from "./helpers/convert";
+
+describe("locals", () => {
+  testCase.test(
+    "locals",
+    `
+    locals {
+      service_name = "forum"
+      owner        = "Community Team"
+      is_it_great  = true
+      how_many     = 42
+    }
+    output "combined-so-it-does-not-get-removed" {
+      value = "\${local.service_name},\${local.owner},\${local.is_it_great},\${local.how_many}"
+    }
+    `,
+    [],
+    Synth.yes
+  );
+
+  testCase.test(
+    "multiple locals blocks",
+    `
+    locals {
+      service_name = "forum"
+      owner        = "Community Team"
+    }
+
+    locals {
+      is_it_great  = true
+      how_many     = 42
+    }
+    
+    output "combined-so-it-does-not-get-removed" {
+      value = "\${local.service_name},\${local.owner},\${local.is_it_great},\${local.how_many}"
+    }`,
+    [],
+    Synth.yes
+  );
+});
